requirements.txt
pandas==2.2.3
scikit-learn==1.5.2
numpy==2.1.3
langchain==1.0.5
langchain-community==0.4.1
langchain-google-genai==3.0.2
langchain-huggingface==0.0.3
langchain-chroma==0.0.4
sentence-transformers==5.1.2
chromadb==1.3.4
fastapi==0.115.5
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
streamlit==1.39.0
requests==2.32.5
